Handle Windows path mode suffixes
